Remove the 'rules' dependency on Makefile, and remove the Makefile.
There was quite a bit of cruft in Makefile, distributed responsibility
between rules and Makefile, and some contention on installed files.
Now there is only one makefile, and package definition is limited to
the /debian directory.

The update-piuparts-*-setup scripts are modified to use 'rules' as the
makefile, though it doesn't work yet, and looks like it might not
ever.

This is squashed to a single commit, since there was necessarily
considerable breakage in intermediate steps.